Поддержка Darwin, разделение функциональности между UNIX системами
